+++ .github/workflows/build.yml
@@ -1,0 +1,40 @@
+name: Build
+
+on:
+  push:
+  workflow_dispatch:
+
+jobs:
+  build:
+    name: Build
+    runs-on: ubuntu-latest
+    steps:
+
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Setup Java
+        uses: actions/setup-java@v4
+        with:
+          distribution: 'temurin'
+          java-version: 21
+
+      - name: Setup Gradle
+        uses: gradle/actions/setup-gradle@v4
+
+      - name: Set environments
+        run: |
+          {
+            echo "version=v$(grep versionName app/build.gradle | awk -F\" '{print $2}')"
+            echo "commit=$(echo ${{ github.sha }} | cut -c -7)"
+            echo "repo=$(echo ${GITHUB_REPOSITORY#$GITHUB_REPOSITORY_OWNER/})"
+          } >> $GITHUB_ENV
+
+      - name: Build
+        run: ./gradlew assembleDebug --no-daemon --stacktrace --warning-mode=all
+
+      - name: Upload
+        uses: actions/upload-artifact@v4
+        with:
+          name: ${{ env.repo }}(${{ env.version }}@${{ env.commit }})
+          path: app/build/outputs/apk/debug/app-debug.apk
