+++ README.md
@@ -1,46 +1,32 @@
+# Image2Ascii
 
-# 画像をASCIIアートに変換するAndroid用アプリ(image to ASCII)
-
-このAndroidアプリは画像をASCIIアートに変換する事が出来ます。
+このAndroidアプリは、任意の画像をASCIIアートに変換できます
 
 ## 機能
 
-- デバイスのストレージから画像を選択。
-- 選択した画像をASCIIアートに変換。
-- 生成されたASCIIアートを `.txt` ファイルとして内部ストレージ(/sdcard/Android/data/com.image2ascii/files/内に保存されます)
-- アプリ内で直接ASCIIアートを表示。(端末によって表示結果レイアウトが崩れるが生成されたtxtに問題はない)
+- デバイスのストレージから画像を選択し、画像をASCIIアートに変換
+- 生成されたASCIIアートを `.txt` ファイルとして  
+  内部ストレージ(`/sdcard/Android/data/com.image2ascii/files/`)に保存
+- アプリ内で直接ASCIIアートを表示  
+  ※端末の画面密度によってはレイアウトが崩れます
 
 ## 使用方法
 
 1. **権限を付与**:
-   - 初回起動時にアプリはストレージへのアクセス許可を求めますので
-　　　権限を許可してください。
+  - 初回起動時にアプリはストレージへのアクセス許可を要求しますので、権限を許可してください。
 
 2. **画像を選択**:
-   - 「画像を選択」ボタンをタップしてストレージを開きます。
-   - ストレージから画像ファイルを選択します。
+  - ｢<kbd>画像を選択</kbd>｣ボタンをタップしてストレージを開きます。
+  - ストレージから画像ファイルを選択します。
 
 3. **ASCIIアートに変換**:
-   - 画像を選択すると、「ASCIIに変換」ボタンが有効になります。
-   - ボタンをタップして変換を開始します。
-   - ASCIIアートがアプリ内に表示されます。
+  - 画像を選択すると、｢<kbd>ASCIIに変換</kbd>｣ボタンが有効になります。
+  - ボタンをタップして変換を開始します。
+  - ASCIIアートがアプリ内に表示されます。
 
 4. **ASCIIアートを保存**:
-   - ASCIIアートは自動的にアプリの/sdcard/Android/data/com.image2ascii/files/内に `.txt` ファイルとして保存されます。
-   - 保存されたファイルの場所は通知メッセージとして表示されます。
-
-## 使用ライブラリ
-
-このアプリは以下のライブラリを使用しています：
-
-- **AndroidX ライブラリ**:
-  - `androidx.appcompat:appcompat:1.7.0`
-  - `androidx.core:core:1.12.0`
-  - `androidx.constraintlayout:constraintlayout:2.2.0`
-  - `androidx.activity:activity-ktx:1.8.0`
-
-このアプリの参考元ソースは下記です
-https://tool-taro.com/blog/java/230/
+  - ASCIIアートは自動的に`/sdcard/Android/data/com.image2ascii/files/` に `.txt` ファイルとして保存されます。
+  - 保存されたファイルの場所は通知メッセージとして表示されます。
 
 ## 権限
 
@@ -56,14 +42,10 @@ https://tool-taro.com/blog/java/230/
 - `.png`
 - `.bmp`
 
-##　※注意事項と補足
-端末によっては変換後の表示文字の
-レイアウトが崩れますが
-ストレージ内に保存されたtxtをPCなどで開けば正常に生成されてるのが確認出来ます。
+## 注意事項
+タブレット端末 (1920x1200, 240dpi) で行って調整したので、スマホではレイアウトが崩れる可能性があります。
+生成された `.txt` ファイルをパソコン等の端末で開けば正常に表示出来ます。
 
-また、このアプリは更新予定がありませんので
-自由にカスタマイズしてお使い下さい。
+## 参考
 
-## 動作確認端末
-タブレット端末で行って調整したのでスマホでは特にレイアウトが崩れます。
-生成されたtxtファイルをPCで開けば正常に表示出来ます。
+- [**画像→AA**](https://tool-taro.com/blog/java/230/)
