+++ app/proguard-rules.pro
@@ -1,0 +1,2 @@
+-dontobfuscate
+-keepparameternames
