+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,13 @@
 <resources>
-    <string name="app_name">Image to ASCII</string>
+    <string name="app_name">Image2Ascii</string>
+
+    <string name="cannot_get_image_path">画像のパスを取得できません</string>
+    <string name="cannot_load_image">画像を読み込めません</string>
+    <string name="convert_ascii_art">アスキーアートに変換</string>
+    <string name="file_save_failed">ファイルの保存に失敗しました</string>
+    <string name="file_save_success">ファイルが保存されました:</string>
+    <string name="need_storage_permission">ストレージアクセス権限が必要です</string>
+    <string name="select_image">画像を選択</string>
+    <string name="selected_image">画像を選択しました</string>
+    <string name="storage_permission_granted">ストレージアクセス権限が許可されました</string>
 </resources>
