--- app/src/main/res/values/styles.xml
@@ -1,9 +1,0 @@
-<resources>
-
-    <!-- ベーステーマ -->
-    <style name="Theme.AppCompat.Light.NoActionBar" parent="Theme.AppCompat.Light">
-        <item name="windowActionBar">false</item>
-        <item name="windowNoTitle">true</item>
-    </style>
-
-</resources>
