+++ gradle.properties
@@ -1,4 +1,2 @@
-android.useAndroidX=true
-android.enableJetifier=true
 org.gradle.configuration-cache=true
 org.gradle.jvmargs=-Xmx4096m -Dfile.encoding=UTF-8
