+++ settings.gradle
@@ -4,7 +4,7 @@ pluginManagement {
         mavenCentral()
     }
     plugins {
-        id 'com.android.application' version '8.7.3' apply false
+        id 'com.android.application' version '8.8.0' apply false
     }
 }
 
@@ -14,5 +14,6 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "image to ASCII"
+
+rootProject.name = "Image2Ascii"
 include 'app'
